+++ src/js/entities-service/entities/flows.js
@@ -26,7 +26,6 @@ const _Model = BaseModel.extend({
   },
   getOwner() {
     const owner = this.get('_owner');
-    if (!owner) return;
     return Radio.request('entities', `${ owner.type }:model`, owner.id);
   },
   isDone() {
